fix: unicode filenames on Linux
This bug caused CJK chars in filename to be stripped out on Linux.
For example, `/media/mydisk/fooにほん.rar`  would result in
`/media/mydisk/foo` so the open will fail

due to

```cpp
    if (r->ArcNameW!=nullptr && *r->ArcNameW!=0)
      ArcName=r->ArcNameW;
    else
      CharToWide(AnsiArcName,ArcName);

    Data->Arc.Open(ArcName,FMF_OPENSHARED)
```

in `Data->Arc.Open`, ArcName is converted with `WideToChar`
causing the CJK bytes to get lost. So we pass it as AnsiArcName,
so `CharToWide` is called first, causing to conversion to work.

Co-authored-by: Danyel <vcs@danyel.io>
Fixes #44